Manually fit and pool Cox proportional hazards model results from multiple imputed datasets

Function manually fits and pools results from Cox proportional hazards models using a list of imputed datasets.

Usage

cox_pooling(
  x,
  surv_formula = stats::as.formula(survival::Surv(fu_itt_months, death_itt) ~ treat)
)

Arguments

x

list of imputed datasets with weights or raking weights (raking_weights) and cluster (matched datasets only)

surv_formula

formula for the Cox proportional hazards model (default is Surv(fu_itt_months, death_itt) ~ treat)

Value

list of data frames with updated raking weights (raking_weights)

Details

The function requires an list of imputed data frames with weights and cluster (matching) information and a formula for the Cox proportional hazards model. The data frames must have a column names weights and subclass (for matched datasets to indicate the cluster membership).

This is a convenience wrapper for fitting Cox models to multiple imputed datasets which do not come as a mimids or wimids object. This is useful when there are intermediate steps in the analysis pipeline, such as computing raking weights via raking_weights which are so far not straightforward to implement in a mimids or wimids object.

The function follows the following logic:

Fit a Cox proportional hazards model to each imputed dataset; if a subclass column is present in the data, it is used as a cluster variable for matched pairs

Pool the results using the pool.scalar function from the mice package, that is, univariate estimates (qbar) are pooled via formula (3.1.2) Rubin (1987) and the total variance (t) is estimated via robust standardard errors according to formula (3.1.5) Rubin (1987).

Compute the confidence intervals and exponentiate the results via exp(qbar +/- 1.96 * sqrt(t))

See also

Examples

library(encore.analytics)
library(mice)
#> 
#> Attaching package: ‘mice’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following objects are masked from ‘package:base’:
#> 
#>     cbind, rbind
library(MatchThem)
#> 
#> Attaching package: ‘MatchThem’
#> The following objects are masked from ‘package:mice’:
#> 
#>     cbind, pool
#> The following object is masked from ‘package:base’:
#> 
#>     cbind

# simulate a cohort with 1,000 patients with 20% missing data
data <- simulate_data(
 n = 500,
 imposeNA = TRUE,
 propNA = 0.2
 )

# impute the data
set.seed(42)
mids <- mice(data, m = 5, print = FALSE)
#> Warning: Number of logged events: 801

# fit a propensity score model
fit <- as.formula(treat ~ dem_age_index_cont + dem_sex_cont + c_smoking_history)

# weight (or alternatively match) patients within each imputed dataset
wimids <- weightthem(
 formula = fit,
 datasets = mids,
 approach = "within",
 method = "glm",
 estimand = "ATO"
 )
#> Estimating weights     | dataset: #1
#>  #2
#>  #3
#>  #4
#>  #5
#> 

# create a list of imputed and weighted datasets
wimids_list <- MatchThem::complete(wimids, action = "all", all = FALSE, include = FALSE)

# fit a survival model
cox_fit <- as.formula(survival::Surv(fu_itt_months, death_itt) ~ treat)

# fit and pool Cox proportional hazards model results
cox_pooling(wimids_list, surv_formula = cox_fit)
#>    term  estimate  std.error statistic      p.value  conf.low conf.high
#> 1 treat 0.7334003 0.08794506 -3.525651 0.0004623598 0.6170148 0.8717392
#>              b       df dfcom        fmi      lambda m         riv        ubar
#> 1 4.813004e-05 486.9941   496 0.01151869 0.007467489 5 0.007523672 0.007676578